+++ NEWS.md
@@ -1,5 +1,8 @@
 # r4ct 0.2.5
-change the name to `r4ct`
+
+* change the name to `r4ct`
+* add functions for Farrington-Manning test method for non-inferiority test in
+the context of comparing two proportions
 
 # shinyapps4clinicaltrial 0.2.4